requirements.txt
fcm-django==0.3.7
django-admin-interface==0.16.1
django==3.2
django-dotenv==1.4.2
python-dotenv==0.20.0
psycopg2-binary==2.9.3
djangorestframework==3.13.1
